==> notebooks/0. ir-datasets.ipynb <==
# !pip install --upgrade ir_datasets

from pathlib import Path

import pandas as pd
import numpy as np

import ir_datasets

import config

dataset_name = config.DATASET
dataset = ir_datasets.load(dataset_name)

dataset.docs_count()

n_docs = 1000

for i, data in enumerate(zip(dataset.queries_iter(), dataset.docs_iter(), dataset.qrels_iter())):
    query, doc, rel = data
    print(query)
    print(doc)
    print(rel)
    break

df_queries = pd.DataFrame(dataset.queries_iter())
print(len(df_queries))
df_queries.head()

df_docs = pd.DataFrame(dataset.docs_iter())
print(len(df_docs))
df_docs.head()

df_qrels = pd.DataFrame(dataset.qrels_iter())
print(len(df_qrels))
df_qrels.head()

df_qrels.relevance.value_counts()

df_docs = df_docs.rename({"text":"doc_text"}, axis=1).drop(["metadata", "title"], axis=1)
df_docs.head()

df_queries = df_queries.rename({"text":"query_text"}, axis=1).drop(["metadata"], axis=1)
df_queries.head()

df = df_qrels.merge(df_docs).merge(df_queries)[["query_text", "doc_text", "relevance"]]
print(len(df))
df.head()

df.isna().sum()

df.relevance.value_counts()

df_negative = df.copy()

np.random.shuffle(df_negative.doc_text)
df_negative["relevance"] = 0
df_negative.head()

df = pd.concat([df, df_negative])
print(len(df))
df.head()

save_path = Path(f"data/{dataset_name}")
save_path.mkdir(exist_ok=True, parents=True)
df.to_pickle(save_path/"data.pkl")
