==> notebooks/1. faster_cosine.ipynb <==
import copy 
from typing import Optional

import torch
from torch import nn
from torch import Tensor, device

import numpy as np
from torch.utils.data import DataLoader
from tqdm.notebook import tqdm

torch.manual_seed(0)

## Faster and batched cosine similarity

n_emb = 10000 #bigger than this will cause RAM and GPU memory issues
emb_size = 768
a = torch.rand(n_emb, emb_size)
a.dtype, a.shape

def original_cos_sim(a: Tensor, b: Tensor):
    """
    Computes the cosine similarity cos_sim(a[i], b[j]) for all i and j.
    :return: Matrix with res[i][j]  = cos_sim(a[i], b[j])
    """
    if not isinstance(a, torch.Tensor):
        a = torch.tensor(a)

    if not isinstance(b, torch.Tensor):
        b = torch.tensor(b)

    if len(a.shape) == 1:
        a = a.unsqueeze(0)

    if len(b.shape) == 1:
        b = b.unsqueeze(0)

    a_norm = torch.nn.functional.normalize(a, p=2, dim=1)
    b_norm = torch.nn.functional.normalize(b, p=2, dim=1)
    return torch.mm(a_norm, b_norm.transpose(0, 1))

from pynvml import *

def get_size_of_tensor(a):
    return a.element_size() * a.nelement()

def get_free_gpu_memory():
    nvmlInit()
    h = nvmlDeviceGetHandleByIndex(0)
    info = nvmlDeviceGetMemoryInfo(h)
    # print(f'total    : {info.total/1e6}MB')
    # print(f'free     : {info.free/1e6}MB')
    # print(f'used     : {info.used/1e6}MB')
    return info.free

get_free_gpu_memory()/1e6

def suggest_batch_size(a, b, fp16):
    import sys

    dim = a.shape[-1]
    single_emb = torch.rand(1, dim)
    size_of_emb = get_size_of_tensor(single_emb)

    if fp16:
        size_of_emb = size_of_emb/2

    free_gpu_mem = get_free_gpu_memory()
    size_of_float = 4
    safety_factor = 2
    #assume a is batched
    # (free_mem - size_a - size_b)*batches = len(a)*len(b)*size_of_float
    # (free_mem - size_a - size_b)/size_of_float/len(b) = len(a)/batches = batch_size

    batch_size = (free_gpu_mem - len(a)*size_of_emb - len(b)*size_of_emb)/size_of_float/len(b) 
    batch_size = int(min(len(a), batch_size)/8)*8/safety_factor
    return batch_size

suggest_batch_size(a, a, fp16=False)

def get_tensor_for_cos_sim(a: Tensor, compute_device: Optional[str]=None, fp16:bool=False, normalize:bool=True):
    if not isinstance(a, torch.Tensor):
        a = torch.tensor(a)

    if len(a.shape) == 1:
        a = a.unsqueeze(0)

    if compute_device:
        a = a.to(compute_device)

    if fp16:
        a = a.type(torch.float16)

    return torch.nn.functional.normalize(a, p=2, dim=1) if normalize else a


def cos_sim(a: Tensor, b: Optional[Tensor]=None, compute_device: Optional[str]=None, return_device: str="cpu", fp16:bool=False, batch_size="auto", normalize:bool=True):
    """
    Computes the cosine similarity cos_sim(a[i], b[j]) for all i and j.
    :return: Matrix with res[i][j]  = cos_sim(a[i], b[j])
    """
    a = get_tensor_for_cos_sim(a, compute_device, fp16, normalize)

    if b is None:
        b = a
    else:
        b = get_tensor_for_cos_sim(b, compute_device, fp16)

    sim = []
    if batch_size == "auto":
        batch_size = suggest_batch_size(a, b, fp16=fp16)
    chunks = int(len(a)/batch_size)
    for chunk in torch.chunk(a, chunks):
        sim_batch = torch.mm(chunk, b.transpose(0, 1))
        sim.append(sim_batch.to("cpu"))
    sim = torch.cat(sim)

    del chunk; del sim_batch
    del a; del b

    if fp16:
      sim = sim.type(torch.float32)

    if return_device:
      sim = sim.to(return_device)
      
    return sim

%%timeit
original_cos_sim(a, a)

%timeit x = cos_sim(a, compute_device=None, fp16=False)

%timeit x = cos_sim(a, a, compute_device=None, fp16=False)

%timeit y = cos_sim(a, a, compute_device="cuda", fp16=False)

%timeit y = cos_sim(a, compute_device="cuda", fp16=False)

%timeit z = cos_sim(a, a, compute_device="cuda", fp16=True)

x = cos_sim(a, a, compute_device=None, fp16=False)
y = cos_sim(a, a, compute_device="cuda", fp16=False)
z = cos_sim(a, a, compute_device="cuda", fp16=True)

torch.all(torch.isclose(x,y))

torch.all(torch.isclose(x,z, atol=1e-06)) #cosine sim upto 6 decimal places

torch.all(torch.isclose(x,z, atol=1e-03)) #cosine sim upto 3 decimal places is same which is good enough

torch.cuda.empty_cache()

!nvidia-smi

cos_sim(a, a, compute_device="cuda")

!nvidia-smi

a.shape

z = cos_sim(a, a, compute_device="cpu", fp16=False, return_device=None)
z.shape

z[:2, :10]

batch_size = len(a)/5
z = cos_sim(a, a, compute_device="cpu", fp16=False, return_device=None, batch_size=batch_size)
z.shape

z[:2, :10]

z = cos_sim(a, a, compute_device="cuda", fp16=True, return_device='cpu', batch_size=batch_size)
z.shape

z[:2, :10]

z = original_cos_sim(a, a)
z.shape

z[:2, :10]
