==> notebooks/2. clip_layers.ipynb <==
import copy 
from pathlib import Path
from typing import Optional

import torch
from torch import nn
from torch import Tensor, device

import pandas as pd
import numpy as np

from tqdm.notebook import tqdm, trange
from fastprogress.fastprogress import master_bar, progress_bar

from sentence_transformers import SentenceTransformer, util

import config
from faiss_indexer import FAISS

config.MODEL

distill_model = SentenceTransformer(config.MODEL_SMALL)
config.MODEL_SMALL

model = SentenceTransformer(config.MODEL)
LAYERS = 12

dim = model.encode(["hello"]).shape[-1]
dim

n_emb = 10000
emb_size = dim
a = torch.rand(n_emb, emb_size)
a.dtype, a.shape

## Prune top layers from sentence-transformer

# encoder = model._modules['0']._modules['auto_model']._modules['encoder']._modules['layer']
# encoder

def deleteEncodingLayers(model, layers_to_keep):  # must pass in the full model
    oldModuleList = model._modules['0']._modules['auto_model']._modules['encoder']._modules['layer']
    newModuleList = nn.ModuleList()

    # Now iterate over all layers, only keepign only the relevant layers.
    for i in layers_to_keep:
        newModuleList.append(oldModuleList[i])

    # create a copy of the model, modify it with the new list, and return
    copyOfModel = copy.deepcopy(model)
    copyOfModel._modules['0']._modules['auto_model']._modules['encoder']._modules['layer'] = newModuleList

    return copyOfModel

for remove_n_layers in range(3):
  layers_to_keep = list(range(0, LAYERS-remove_n_layers))
  print(layers_to_keep)
  small_model = deleteEncodingLayers(model, layers_to_keep)
  new_encodings = small_model.encode('hello')
  encodings = model.encode('hello')
  cos = util.cos_sim(encodings, new_encodings)
  print(cos)

## Check time reduction

%timeit model.encode('hello how are you?')

%timeit small_model.encode('hello how are you?')

## Check quality

dataset_name = config.DATASET
save_path = Path(f"data/{dataset_name}")
df = pd.read_pickle(save_path/"data.pkl")
df.head()

texts = df.query_text.unique().tolist()

vectors = model.encode(texts)
small_vectors = small_model.encode(texts)
distill_vectors = distill_model.encode(texts)
vectors.shape, small_vectors.shape, distill_vectors.shape

index = FAISS(768, gpu=False)
small_index = FAISS(768, gpu=False)
distill_index = FAISS(384, gpu=False)

for i in trange(len(texts)):
    index.add(texts[i], [vectors[i]])
    small_index.add(texts[i], [small_vectors[i]])
    distill_index.add(texts[i], [distill_vectors[i]])

## paraphrase-mpnet-base-v2 results

query = "how to get better health"

query_vector = model.encode([query])
print(query_vector.shape)
index.search(query_vector, dataframe=True, k=20)

## cropped paraphrase-mpnet-base-v2 results

query_vector = small_model.encode([query])
print(query_vector.shape)
small_index.search(query_vector, dataframe=True, k=20)

## paraphrase-MiniLM-L3-v2 results

query_vector = distill_model.encode([query])
print(query_vector.shape)
distill_index.search(query_vector, dataframe=True, k=20)
